improve many arg api testing
