+++ examples/api-arg-many/index.rsh
@@ -1,19 +1,22 @@
 'reach 0.1'
 
-export const main = Reach.App(() => {
+const manyArgFunSig = Fun([UInt, UInt, UInt, UInt,
+                           UInt, UInt, UInt, UInt,
+                           UInt, UInt, UInt, UInt,
+                           UInt, UInt, UInt, UInt,
+                           UInt, UInt, UInt, UInt,
+                          ],
+                          UInt)
+
+export const mainServer = Reach.App(() => {
   const Deployer = Participant('Deployer', {
     ready: Fun([], Null),
   })
-  const A = API('A', {
-    sumMany: Fun([UInt, UInt, UInt, UInt,
-                  UInt, UInt, UInt, UInt,
-                  UInt, UInt, UInt, UInt,
-                  UInt, UInt, UInt, UInt,
-                  //// I wanted to test more than 15, but when 16 worked I
-                  //// decided to keep going higher.
-                  UInt, UInt, UInt, UInt,
-                 ],
-                 UInt)
+  const A = API({
+    sumMany: manyArgFunSig,
+  })
+  const V = View({
+    vSumMany: manyArgFunSig,
   })
   setOptions({
     connectors: [
@@ -24,25 +27,71 @@ export const main = Reach.App(() => {
   })
   init()
 
+  // TODO - when I wrote this as (...args) => ..., I got an error when setting it to a view that expected a certain number of arguments.
+  const sum = (v0,v1,v2,v3,
+               v4,v5,v6,v7,
+               v8,v9,v10,v11,
+               v12,v13,v14,v15,
+               v16,v17,v18,v19
+              ) => {
+                return v0 + v1 + v2 + v3
+                  + v4 + v5 + v6 + v7
+                  + v8 + v9 + v10 + v11
+                  + v12 + v13 + v14 + v15
+                  + v16 + v17 + v18 + v19
+              }
+
   Deployer.publish()
+  V.vSumMany.set(sum)
   Deployer.interact.ready()
   commit()
 
-  const [args, sumManyRet] = call(A.sumMany)
 
-  // FIXME - Reach doesn't yet support turning a DL Tuple into an array.
-  //const asArray = array(UInt, args)
-  const asArray = array(UInt, [args[0], args[1], args[2], args[3],
-                               args[4], args[5], args[6], args[7],
-                               args[8], args[9], args[10], args[11],
-                               args[12], args[13], args[14], args[15],
-                               args[16], args[17], args[18], args[19],
-                              ])
-  const sum = asArray.reduce(0, (accum, x) => accum + x)
-  sumManyRet(sum)
+  const [args, sumManyRet] = call(A.sumMany)
+  sumManyRet(sum(...args))
+  commit()
+
+  // Needs a last call to delete the program.
+  Deployer.publish()
+  commit()
+
+})
+
+export const mainClient = Reach.App(() => {
+  const Deployer = Participant('Deployer', {
+    serverCtcInfo: Contract,
+    ready: Fun([], Null),
+  })
+  const A = API({
+    poke: manyArgFunSig,
+  })
+  setOptions({
+    connectors: [
+      // With 16 arguments, the Solidity compiler errors saying that the stack is too deep.
+      //ETH,
+      ALGO,
+    ]
+  })
+  init()
+
+  Deployer.only(() => {
+    const serverCtcInfo = declassify(interact.serverCtcInfo)
+  })
+
+  Deployer.publish(serverCtcInfo)
+  Deployer.interact.ready()
+
+  const sumServer = remote(serverCtcInfo, {
+    sumMany: manyArgFunSig,
+  })
 
   commit()
-  Deployer.publish()
+
+  const [args, pokeRet] = call(A.poke)
+
+  const sum = sumServer.sumMany(...args)
+
+  pokeRet(sum)
 
   commit()
